pyproject.toml
[tool.poetry]
name = "scrapybara-playground"
version = "0.1.0"
description = ""
authors = ["Justin Sun <justin@scrapybara.com>", "Nalin Semwal <nalin@scrapybara.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
scrapybara = "2.0.4"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
fastapi = "^0.115.5"
python-dotenv = "^1.0.1"
websocket = "^0.2.1"
supabase = "^2.10.0"
anthropic = "^0.39.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
